// components/manage/newpost/block-editor/enums/button-type.enum.ts
export enum ButtonStyle {
  bold = 'bold',
  italic = 'italic',
  underline = 'underline',
  strikeThrough = 'strikeThrough',
  insertOrderedList = 'insertOrderedList',
  insertUnorderedList = 'insertUnorderedList',
  justifyLeft = 'justifyLeft',
  justifyCenter = 'justifyCenter',
  justifyRight = 'justifyRight',
  subscript = 'subscript',
  superscript = 'superscript',
}